+++ src/main/kotlin/com/oneinch/repository/dao/FakeTokenQuoteEntity.kt
@@ -7,9 +7,9 @@ import javax.persistence.Entity
 import javax.persistence.Id
 
 @Entity
-class FakeTokenQuoteEntity(var chainId: Int, var symbol: String, @Id var address: String, var readable: Double, var origin: String, var decimals: Int) {
+class FakeTokenQuoteEntity(var chainId: Int, var symbol: String, @Id var address: String, var readable: Double, var origin: String, var decimals: BigInteger) {
 
-    constructor() : this(0, "", "", 0.0, "0", 0)
+    constructor() : this(0, "", "", 0.0, "0", BigInteger.valueOf(0))
 }
 
 fun FakeTokenQuoteEntity.toTokenQuote(): TokenQuote {
